+++ src/fapilog/testing/mock_sinks.py
@@ -5,7 +5,7 @@ import random
 import time
 from typing import Any, Dict, List
 
-from .._internal.queue import Sink
+from ..sinks import Sink
 
 
 class RecordingSink(Sink):
