+++ tests/test_error_handling.py
@@ -495,7 +495,8 @@ class TestComprehensiveErrorHandling:
 
     def test_queue_error_handling_with_multiple_sinks(self):
         """Test queue error handling when multiple sinks fail."""
-        from fapilog._internal.queue import QueueWorker, Sink
+        from fapilog._internal.queue_worker import QueueWorker
+        from fapilog.sinks import Sink
 
         class FailingSink(Sink):
             async def write(self, event_dict):
