+++ tests/test_integration_simple.py
@@ -6,8 +6,8 @@ from unittest.mock import Mock, patch
 
 import pytest
 
-from fapilog._internal.queue import Sink
 from fapilog._internal.sink_registry import SinkRegistry
+from fapilog.sinks import Sink
 from fapilog.testing.integration import SinkIntegrationTester
