+++ tests/test_performance_testing.py
@@ -5,7 +5,7 @@ from unittest.mock import Mock, patch
 
 import pytest
 
-from fapilog._internal.queue import Sink
+from fapilog.sinks import Sink
 from fapilog.testing.performance import SinkPerformanceTester
